requirements.txt
Django>=4.0,<4.1
wagtail>=2.16,<2.17
django-extensions==3.1.5
graphene==2.1.9
graphene-django==2.13.0
graphql-core==2.3.2
wagtail-grapple==0.14.1
wagtail-headless-preview==0.2.1
